--- src/Currency/data/iso-currencies.ts
@@ -1,1290 +1,0 @@
-export default {
-  AED: {
-    code: 'AED',
-    symbol: 'د.إ.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  AFN: {
-    code: 'AFN',
-    symbol: '؋',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  ALL: {
-    code: 'ALL',
-    symbol: 'Lek',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  AMD: {
-    code: 'AMD',
-    symbol: '֏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  ANG: {
-    code: 'ANG',
-    symbol: 'ƒ',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  AOA: {
-    code: 'AOA',
-    symbol: 'Kz',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  ARS: {
-    code: 'ARS',
-    symbol: '$',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  AUD: {
-    code: 'AUD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  AWG: {
-    code: 'AWG',
-    symbol: 'ƒ',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  AZN: {
-    code: 'AZN',
-    symbol: '₼',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  BAM: {
-    code: 'BAM',
-    symbol: 'КМ',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  BBD: {
-    code: 'BBD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  BDT: {
-    code: 'BDT',
-    symbol: '৳',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%s %ns%v',
-  },
-  BGN: {
-    code: 'BGN',
-    symbol: 'лв.',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  BHD: {
-    code: 'BHD',
-    symbol: 'د.ب.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%s %ns%v',
-  },
-  BIF: {
-    code: 'BIF',
-    symbol: 'FBu',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  BMD: {
-    code: 'BMD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  BND: {
-    code: 'BND',
-    symbol: '$',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  BOB: {
-    code: 'BOB',
-    symbol: 'Bs',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  BRL: {
-    code: 'BRL',
-    symbol: 'R$',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  BSD: {
-    code: 'BSD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  BTC: {
-    code: 'BTC',
-    symbol: 'Ƀ',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  BTN: {
-    code: 'BTN',
-    symbol: 'Nu.',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 1,
-    pattern: '%s %ns%v',
-  },
-  BWP: {
-    code: 'BWP',
-    symbol: 'P',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  BYR: {
-    code: 'BYR',
-    symbol: 'р.',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  BZD: {
-    code: 'BZD',
-    symbol: 'BZ$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CAD: {
-    code: 'CAD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CDF: {
-    code: 'CDF',
-    symbol: 'FC',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  CHF: {
-    code: 'CHF',
-    symbol: 'CHF',
-    thousandsSeparator: '\'',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  CLP: {
-    code: 'CLP',
-    symbol: '$',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  CNY: {
-    code: 'CNY',
-    symbol: '¥',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  COP: {
-    code: 'COP',
-    symbol: '$',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  CRC: {
-    code: 'CRC',
-    symbol: '₡',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CUC: {
-    code: 'CUC',
-    symbol: 'CUC',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CUP: {
-    code: 'CUP',
-    symbol: '$MN',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CVE: {
-    code: 'CVE',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  CZK: {
-    code: 'CZK',
-    symbol: 'Kč',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  DJF: {
-    code: 'DJF',
-    symbol: 'Fdj',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  DKK: {
-    code: 'DKK',
-    symbol: 'kr.',
-    thousandsSeparator: '',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  DOP: {
-    code: 'DOP',
-    symbol: 'RD$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  DZD: {
-    code: 'DZD',
-    symbol: 'د.ج.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  EGP: {
-    code: 'EGP',
-    symbol: 'ج.م.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  ERN: {
-    code: 'ERN',
-    symbol: 'Nfk',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  ETB: {
-    code: 'ETB',
-    symbol: 'ETB',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  EUR: {
-    code: 'EUR',
-    symbol: '€',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  FJD: {
-    code: 'FJD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  FKP: {
-    code: 'FKP',
-    symbol: '£',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  GBP: {
-    code: 'GBP',
-    symbol: '£',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  GEL: {
-    code: 'GEL',
-    symbol: 'Lari',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  GHS: {
-    code: 'GHS',
-    symbol: '₵',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  GIP: {
-    code: 'GIP',
-    symbol: '£',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  GMD: {
-    code: 'GMD',
-    symbol: 'D',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  GNF: {
-    code: 'GNF',
-    symbol: 'FG',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  GTQ: {
-    code: 'GTQ',
-    symbol: 'Q',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  GYD: {
-    code: 'GYD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  HKD: {
-    code: 'HKD',
-    symbol: 'HK$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  HNL: {
-    code: 'HNL',
-    symbol: 'L.',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  HRK: {
-    code: 'HRK',
-    symbol: 'kn',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  HTG: {
-    code: 'HTG',
-    symbol: 'G',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  HUF: {
-    code: 'HUF',
-    symbol: 'Ft',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  IDR: {
-    code: 'IDR',
-    symbol: 'Rp',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  ILS: {
-    code: 'ILS',
-    symbol: '₪',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  INR: {
-    code: 'INR',
-    symbol: '₹',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  IQD: {
-    code: 'IQD',
-    symbol: 'د.ع.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  IRR: {
-    code: 'IRR',
-    symbol: '﷼',
-    thousandsSeparator: ',',
-    decimalSeparator: '/',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  ISK: {
-    code: 'ISK',
-    symbol: 'kr.',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 0,
-    pattern: '%ns%v %s',
-  },
-  JMD: {
-    code: 'JMD',
-    symbol: 'J$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  JOD: {
-    code: 'JOD',
-    symbol: 'د.ا.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%s %ns%v',
-  },
-  JPY: {
-    code: 'JPY',
-    symbol: '¥',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  KES: {
-    code: 'KES',
-    symbol: 'KSh',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  KGS: {
-    code: 'KGS',
-    symbol: 'сом',
-    thousandsSeparator: ' ',
-    decimalSeparator: '-',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  KHR: {
-    code: 'KHR',
-    symbol: '៛',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  KMF: {
-    code: 'KMF',
-    symbol: 'CF',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  KPW: {
-    code: 'KPW',
-    symbol: '₩',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  KRW: {
-    code: 'KRW',
-    symbol: '₩',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  KWD: {
-    code: 'KWD',
-    symbol: 'د.ك.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%s %ns%v',
-  },
-  KYD: {
-    code: 'KYD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  KZT: {
-    code: 'KZT',
-    symbol: '₸',
-    thousandsSeparator: ' ',
-    decimalSeparator: '-',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  LAK: {
-    code: 'LAK',
-    symbol: '₭',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  LBP: {
-    code: 'LBP',
-    symbol: 'ل.ل.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  LKR: {
-    code: 'LKR',
-    symbol: '₨',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%s %ns%v',
-  },
-  LRD: {
-    code: 'LRD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  LSL: {
-    code: 'LSL',
-    symbol: 'M',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  LYD: {
-    code: 'LYD',
-    symbol: 'د.ل.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%ns%s%v',
-  },
-  MAD: {
-    code: 'MAD',
-    symbol: 'د.م.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  MDL: {
-    code: 'MDL',
-    symbol: 'lei',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  MGA: {
-    code: 'MGA',
-    symbol: 'Ar',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  MKD: {
-    code: 'MKD',
-    symbol: 'ден.',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  MMK: {
-    code: 'MMK',
-    symbol: 'K',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MNT: {
-    code: 'MNT',
-    symbol: '₮',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MOP: {
-    code: 'MOP',
-    symbol: 'MOP$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MRO: {
-    code: 'MRO',
-    symbol: 'UM',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  MTL: {
-    code: 'MTL',
-    symbol: '₤',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MUR: {
-    code: 'MUR',
-    symbol: '₨',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MVR: {
-    code: 'MVR',
-    symbol: 'MVR',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 1,
-    pattern: '%ns%v %s',
-  },
-  MWK: {
-    code: 'MWK',
-    symbol: 'MK',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MXN: {
-    code: 'MXN',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MYR: {
-    code: 'MYR',
-    symbol: 'RM',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  MZN: {
-    code: 'MZN',
-    symbol: 'MT',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%s%v',
-  },
-  NAD: {
-    code: 'NAD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  NGN: {
-    code: 'NGN',
-    symbol: '₦',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  NIO: {
-    code: 'NIO',
-    symbol: 'C$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  NOK: {
-    code: 'NOK',
-    symbol: 'kr',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  NPR: {
-    code: 'NPR',
-    symbol: '₨',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  NZD: {
-    code: 'NZD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  OMR: {
-    code: 'OMR',
-    symbol: '﷼',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%s %ns%v',
-  },
-  PAB: {
-    code: 'PAB',
-    symbol: 'B/.',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  PEN: {
-    code: 'PEN',
-    symbol: 'S/.',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  PGK: {
-    code: 'PGK',
-    symbol: 'K',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  PHP: {
-    code: 'PHP',
-    symbol: '₱',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  PKR: {
-    code: 'PKR',
-    symbol: '₨',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  PLN: {
-    code: 'PLN',
-    symbol: 'zł',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  PYG: {
-    code: 'PYG',
-    symbol: '₲',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  QAR: {
-    code: 'QAR',
-    symbol: '﷼',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  RON: {
-    code: 'RON',
-    symbol: 'lei',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  RSD: {
-    code: 'RSD',
-    symbol: 'Дин.',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  RUB: {
-    code: 'RUB',
-    symbol: '₽',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  RWF: {
-    code: 'RWF',
-    symbol: 'RWF',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  SAR: {
-    code: 'SAR',
-    symbol: '﷼',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  SBD: {
-    code: 'SBD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SCR: {
-    code: 'SCR',
-    symbol: '₨',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SDD: {
-    code: 'SDD',
-    symbol: 'LSd',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  SDG: {
-    code: 'SDG',
-    symbol: '£‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SEK: {
-    code: 'SEK',
-    symbol: 'kr',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  SGD: {
-    code: 'SGD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SHP: {
-    code: 'SHP',
-    symbol: '£',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SLL: {
-    code: 'SLL',
-    symbol: 'Le',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SOS: {
-    code: 'SOS',
-    symbol: 'S',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SRD: {
-    code: 'SRD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  STD: {
-    code: 'STD',
-    symbol: 'Db',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SVC: {
-    code: 'SVC',
-    symbol: '₡',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  SYP: {
-    code: 'SYP',
-    symbol: '£',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  SZL: {
-    code: 'SZL',
-    symbol: 'E',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  THB: {
-    code: 'THB',
-    symbol: '฿',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  TJS: {
-    code: 'TJS',
-    symbol: 'TJS',
-    thousandsSeparator: ' ',
-    decimalSeparator: ';',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  TMT: {
-    code: 'TMT',
-    symbol: 'm',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  TND: {
-    code: 'TND',
-    symbol: 'د.ت.‏',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 3,
-    pattern: '%s %ns%v',
-  },
-  TOP: {
-    code: 'TOP',
-    symbol: 'T$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  TRY: {
-    code: 'TRY',
-    symbol: 'TL',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  TTD: {
-    code: 'TTD',
-    symbol: 'TT$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  TVD: {
-    code: 'TVD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  TWD: {
-    code: 'TWD',
-    symbol: 'NT$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  TZS: {
-    code: 'TZS',
-    symbol: 'TSh',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  UAH: {
-    code: 'UAH',
-    symbol: '₴',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  UGX: {
-    code: 'UGX',
-    symbol: 'USh',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  USD: {
-    code: 'USD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  UYU: {
-    code: 'UYU',
-    symbol: '$U',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  UZS: {
-    code: 'UZS',
-    symbol: 'сўм',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v %s',
-  },
-  VEB: {
-    code: 'VEB',
-    symbol: 'Bs.',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  VEF: {
-    code: 'VEF',
-    symbol: 'Bs. F.',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  VND: {
-    code: 'VND',
-    symbol: '₫',
-    thousandsSeparator: '.',
-    decimalSeparator: ',',
-    decimalDigits: 1,
-    pattern: '%ns%v %s',
-  },
-  VUV: {
-    code: 'VUV',
-    symbol: 'VT',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 0,
-    pattern: '%ns%v%s',
-  },
-  WST: {
-    code: 'WST',
-    symbol: 'WS$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  XAF: {
-    code: 'XAF',
-    symbol: 'F',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  XCD: {
-    code: 'XCD',
-    symbol: '$',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  XOF: {
-    code: 'XOF',
-    symbol: 'F',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  XPF: {
-    code: 'XPF',
-    symbol: 'F',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%v%s',
-  },
-  YER: {
-    code: 'YER',
-    symbol: '﷼',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%s %ns%v',
-  },
-  ZAR: {
-    code: 'ZAR',
-    symbol: 'R',
-    thousandsSeparator: ' ',
-    decimalSeparator: ',',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  ZMW: {
-    code: 'ZMW',
-    symbol: 'ZK',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-  WON: {
-    code: 'WON',
-    symbol: '₩',
-    thousandsSeparator: ',',
-    decimalSeparator: '.',
-    decimalDigits: 2,
-    pattern: '%ns%s%v',
-  },
-};
