+++ src/CurrencyStore/data.ts
@@ -1,0 +1,10 @@
+import { DeepReadonly } from 'deep-freeze';
+import { CurrencySettings } from '../Currency';
+
+export interface CurrencySettingsInternalStore {
+  [key: string]: DeepReadonly<CurrencySettings>;
+}
+
+const data: CurrencySettingsInternalStore = {};
+
+export default data;
