+++ src/CurrencyStore/index.ts
@@ -1,0 +1,6 @@
+export { default as get } from './get';
+export { default as getAll } from './getAll';
+export { default as register } from './register';
+export { default as registerMultiple } from './registerMultiple';
+export { default as isRegistered } from './isRegistered';
+export { default as deregister } from './deregister';
