+++ src/Money/index.ts
@@ -2,70 +2,15 @@ import BigNumber from 'bignumber.js';
 import deepFreeze from 'deep-freeze';
 import isInt from 'validator/lib/isInt';
 import isFloat from 'validator/lib/isFloat';
-import Currency from '../Currency/Currency';
+import Currency from '../Currency';
 import CurrencyMismatchError from '../errors/CurrencyMismatchError';
 import WrongInputError from '../errors/WrongInputError';
-import Formatter from '../Formatter/Formatter';
-import { CurrencyInputSettings } from '../Currency/CurrencyStore';
-
-/**
- * Rounding modes you can use in your operations.
- * These map directly to
- * [`BigNumber`'s rounding modes](http://mikemcl.github.io/bignumber.js/#constructor-properties).
- *
- * @example
- * Rounding mode usage
- * ```js
- * const
- *      price = new Money('7856', 'USD'), // $78.56
- *      discountedAndRoundedUp = price.multiply('0.70'), // $55.00
- *      discountedAndRoundedDown = price.multiply('0.70', ROUNDING.DOWN); // $54.99
- * ```
- */
-export enum ROUNDING {
-  /**
-   * Rounds away from zero
-   */
-  UP = BigNumber.ROUND_UP,
-  /**
-   * Rounds towards zero
-   */
-  DOWN = BigNumber.ROUND_DOWN,
-  /**
-   * Rounds towards Infinity
-   */
-  CEIL = BigNumber.ROUND_CEIL,
-  /**
-   * Rounds towards Infinity
-   */
-  FLOOR = BigNumber.ROUND_FLOOR,
-  /**
-   * Rounds towards nearest neighbour.
-   * If equidistant, rounds away from zero
-   */
-  HALF_UP = BigNumber.ROUND_HALF_UP,
-  /**
-   * Rounds towards nearest neighbour.
-   * If equidistant, rounds towards zero
-   */
-  HALF_DOWN = BigNumber.ROUND_HALF_DOWN,
-  /**
-   * Rounds towards nearest neighbour.
-   * If equidistant, rounds towards even neighbour
-   */
-  HALF_EVEN = BigNumber.ROUND_HALF_EVEN,
-  /**
-   * Rounds towards nearest neighbour.
-   * If equidistant, rounds towards Infinity
-   */
-  HALF_CEIL = BigNumber.ROUND_HALF_CEIL,
-  /**
-   * Rounds towards nearest neighbour.
-   * If equidistant, rounds towards Infinity
-   */
-  HALF_FLOOR = BigNumber.ROUND_HALF_FLOOR,
-}
-
+import {
+  convertBigNumberToStringInteger,
+  getSmallestUnitAsBigNumber,
+  getSmallestUnitDivisor,
+} from '../utils/internals';
+import ROUNDING from '../constants/ROUNDING';
 
 /**
  * @example
@@ -134,34 +79,6 @@ export default class Money {
   }
 
   /**
-   * Return the absolute monetary value of the current value,
-   * i.e., remove the minus sign if the value is below zero
-   * @returns - new Money instance with the absolute value
-   */
-  absolute() {
-    const newValue = this.value.absoluteValue();
-    return new Money(this.convertBigNumberToStringInteger(newValue), this.currency);
-  }
-
-  /**
-   * Return the largest integer less than or equal to the current value
-   * @returns - new Money instance with the floor value
-   */
-  floor() {
-    const newValue = this.value.decimalPlaces(0, ROUNDING.FLOOR as any);
-    return new Money(this.convertBigNumberToStringInteger(newValue), this.currency);
-  }
-
-  /**
-   * Return the smallest integer greater than or equal to the current value
-   * @returns - new Money instance with the ceiling value
-   */
-  ceil() {
-    const newValue = this.value.decimalPlaces(0, ROUNDING.CEIL as any);
-    return new Money(this.convertBigNumberToStringInteger(newValue), this.currency);
-  }
-
-  /**
    * Checks if the current currency is the same as that of the parameter
    * @param value - value to check currency against the current value; type same as constructor
    * @returns - true if the current value has the same currency as the parameter
@@ -182,15 +99,6 @@ export default class Money {
   }
 
   /**
-   * Format the current value based on the currency
-   * @param settings - formatting settings (optional)
-   * @returns - formatted money
-   */
-  format(settings?: CurrencyInputSettings): string {
-    return this.currency.format(this, settings);
-  }
-
-  /**
    * Get the current value as an instance of BigNumber
    * @returns - Internal BigNumber representation of the current value
    */
@@ -203,7 +111,7 @@ export default class Money {
    * @returns - String integer representation of the current value
    */
   get amountAsStringInteger(): string {
-    return this.convertBigNumberToStringInteger(this.value);
+    return convertBigNumberToStringInteger(this, this.value);
   }
 
   /**
@@ -235,7 +143,7 @@ export default class Money {
    * @returns - new Money instance holding the smallest unit of the current monetary value
    */
   get smallestUnit() {
-    return new Money(this.getSmallestUnitAsBigNumber().toString(), this.currency);
+    return new Money(getSmallestUnitAsBigNumber(this).toString(), this.currency);
   }
 
   /**
@@ -254,13 +162,11 @@ export default class Money {
    * If not, throw an error.
    * @param value - The money object which is used for currency check
    */
-  protected checkValueCurrency(value: Money) {
+  private checkValueCurrency(value: Money) {
     if (!this.hasSameCurrency(value)) {
       throw new CurrencyMismatchError();
     }
   }
-
-
 
   /**
    * Convert the constructor input value to an internal BigNumber instance
@@ -268,13 +174,13 @@ export default class Money {
    * @param BN - BigNumber constructor used by this "Money" instance
    * @returns - Internal BigNumber instance
    */
-  protected preProcessInputValue(value: number|string|Money, BN: typeof BigNumber) {
+  private preProcessInputValue(value: number|string|Money, BN: typeof BigNumber) {
     if (value instanceof Money) {
       this.checkValueCurrency(value);
       return value.amountAsBigNumber;
     }
 
-    const divisor = this.getSmallestUnitDivisor();
+    const divisor = getSmallestUnitDivisor(this);
 
     if (
       divisor.isGreaterThan(1)
@@ -301,38 +207,10 @@ export default class Money {
   }
 
   /**
-   * Get the smallest unit divisor for the current value's currency,
-   * i.e., 10 to the power of the currency's decimal digits.
-   * It is used for converting an integer value to a float value (or vice versa).
-   * @returns - Smallest unit divisor
-   */
-  protected getSmallestUnitDivisor() {
-    const	decimalDigits = this.currency.decimalDigits;
-    return (new BigNumber('10')).exponentiatedBy(decimalDigits);
-  }
-
-  /**
-   * Get the smallest unit of the currency as a big number
-   * @returns - Smallest unit of the currency
-   */
-  protected getSmallestUnitAsBigNumber() {
-    return (new BigNumber('1')).dividedBy(this.getSmallestUnitDivisor());
-  }
-
-  /**
-   * Convert a BigNumber to a string integer
-   * @param value - value to be converted
-   * @return - String integer value of the BigNumber value
-   */
-  protected convertBigNumberToStringInteger(value: BigNumber) {
-    return value.times(this.getSmallestUnitDivisor()).toString();
-  }
-
-  /**
    * Generate a compatible BigNumber constructor
    * @returns - a new BigNumber constructor
    */
-  protected generateBigNumberConstructor() {
+  private generateBigNumberConstructor() {
     return BigNumber.clone({
       DECIMAL_PLACES: 20,
       ROUNDING_MODE: ROUNDING.HALF_UP as any,
@@ -353,16 +231,7 @@ export default class Money {
    * @param currency - currency code as string, instance of `Currency`
    */
   static init(value: number|string|Money, currency: string|Currency) {
-    return new Money(value, currency);
-  }
-
-  /**
-   * Parse a formatted money string into an instance of Money
-   * @param value - the formatted money string
-   * @param settings - the formatting settings
-   * @returns - a Money instance holding the parsed value and currency
-   */
-  static parse(value: string, settings: CurrencyInputSettings|string|Currency): Money {
-    return Formatter.parse(value, settings);
+    const MoneyConstructor = this.constructor as typeof Money;
+    return new MoneyConstructor(value, currency);
   }
 }
