+++ src/constants/ISO_CURRENCIES.ts
@@ -1,0 +1,1449 @@
+export const AED = {
+  code: 'AED',
+  symbol: 'د.إ.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const AFN = {
+  code: 'AFN',
+  symbol: '؋',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const ALL = {
+  code: 'ALL',
+  symbol: 'Lek',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const AMD = {
+  code: 'AMD',
+  symbol: '֏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const ANG = {
+  code: 'ANG',
+  symbol: 'ƒ',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const AOA = {
+  code: 'AOA',
+  symbol: 'Kz',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const ARS = {
+  code: 'ARS',
+  symbol: '$',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const AUD = {
+  code: 'AUD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const AWG = {
+  code: 'AWG',
+  symbol: 'ƒ',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const AZN = {
+  code: 'AZN',
+  symbol: '₼',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const BAM = {
+  code: 'BAM',
+  symbol: 'КМ',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const BBD = {
+  code: 'BBD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const BDT = {
+  code: 'BDT',
+  symbol: '৳',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%s %ns%v',
+};
+
+export const BGN = {
+  code: 'BGN',
+  symbol: 'лв.',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const BHD = {
+  code: 'BHD',
+  symbol: 'د.ب.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%s %ns%v',
+};
+
+export const BIF = {
+  code: 'BIF',
+  symbol: 'FBu',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const BMD = {
+  code: 'BMD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const BND = {
+  code: 'BND',
+  symbol: '$',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const BOB = {
+  code: 'BOB',
+  symbol: 'Bs',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const BRL = {
+  code: 'BRL',
+  symbol: 'R$',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const BSD = {
+  code: 'BSD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const BTC = {
+  code: 'BTC',
+  symbol: 'Ƀ',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const BTN = {
+  code: 'BTN',
+  symbol: 'Nu.',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 1,
+  pattern: '%s %ns%v',
+};
+
+export const BWP = {
+  code: 'BWP',
+  symbol: 'P',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const BYR = {
+  code: 'BYR',
+  symbol: 'р.',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const BZD = {
+  code: 'BZD',
+  symbol: 'BZ$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CAD = {
+  code: 'CAD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CDF = {
+  code: 'CDF',
+  symbol: 'FC',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const CHF = {
+  code: 'CHF',
+  symbol: 'CHF',
+  thousandsSeparator: '\'',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const CLP = {
+  code: 'CLP',
+  symbol: '$',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const CNY = {
+  code: 'CNY',
+  symbol: '¥',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const COP = {
+  code: 'COP',
+  symbol: '$',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const CRC = {
+  code: 'CRC',
+  symbol: '₡',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CUC = {
+  code: 'CUC',
+  symbol: 'CUC',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CUP = {
+  code: 'CUP',
+  symbol: '$MN',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CVE = {
+  code: 'CVE',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const CZK = {
+  code: 'CZK',
+  symbol: 'Kč',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const DJF = {
+  code: 'DJF',
+  symbol: 'Fdj',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const DKK = {
+  code: 'DKK',
+  symbol: 'kr.',
+  thousandsSeparator: '',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const DOP = {
+  code: 'DOP',
+  symbol: 'RD$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const DZD = {
+  code: 'DZD',
+  symbol: 'د.ج.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const EGP = {
+  code: 'EGP',
+  symbol: 'ج.م.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const ERN = {
+  code: 'ERN',
+  symbol: 'Nfk',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const ETB = {
+  code: 'ETB',
+  symbol: 'ETB',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const EUR = {
+  code: 'EUR',
+  symbol: '€',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const FJD = {
+  code: 'FJD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const FKP = {
+  code: 'FKP',
+  symbol: '£',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const GBP = {
+  code: 'GBP',
+  symbol: '£',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const GEL = {
+  code: 'GEL',
+  symbol: 'Lari',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const GHS = {
+  code: 'GHS',
+  symbol: '₵',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const GIP = {
+  code: 'GIP',
+  symbol: '£',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const GMD = {
+  code: 'GMD',
+  symbol: 'D',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const GNF = {
+  code: 'GNF',
+  symbol: 'FG',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const GTQ = {
+  code: 'GTQ',
+  symbol: 'Q',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const GYD = {
+  code: 'GYD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const HKD = {
+  code: 'HKD',
+  symbol: 'HK$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const HNL = {
+  code: 'HNL',
+  symbol: 'L.',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const HRK = {
+  code: 'HRK',
+  symbol: 'kn',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const HTG = {
+  code: 'HTG',
+  symbol: 'G',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const HUF = {
+  code: 'HUF',
+  symbol: 'Ft',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const IDR = {
+  code: 'IDR',
+  symbol: 'Rp',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const ILS = {
+  code: 'ILS',
+  symbol: '₪',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const INR = {
+  code: 'INR',
+  symbol: '₹',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const IQD = {
+  code: 'IQD',
+  symbol: 'د.ع.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const IRR = {
+  code: 'IRR',
+  symbol: '﷼',
+  thousandsSeparator: ',',
+  decimalSeparator: '/',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const ISK = {
+  code: 'ISK',
+  symbol: 'kr.',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 0,
+  pattern: '%ns%v %s',
+};
+
+export const JMD = {
+  code: 'JMD',
+  symbol: 'J$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const JOD = {
+  code: 'JOD',
+  symbol: 'د.ا.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%s %ns%v',
+};
+
+export const JPY = {
+  code: 'JPY',
+  symbol: '¥',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const KES = {
+  code: 'KES',
+  symbol: 'KSh',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const KGS = {
+  code: 'KGS',
+  symbol: 'сом',
+  thousandsSeparator: ' ',
+  decimalSeparator: '-',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const KHR = {
+  code: 'KHR',
+  symbol: '៛',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const KMF = {
+  code: 'KMF',
+  symbol: 'CF',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const KPW = {
+  code: 'KPW',
+  symbol: '₩',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const KRW = {
+  code: 'KRW',
+  symbol: '₩',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const KWD = {
+  code: 'KWD',
+  symbol: 'د.ك.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%s %ns%v',
+};
+
+export const KYD = {
+  code: 'KYD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const KZT = {
+  code: 'KZT',
+  symbol: '₸',
+  thousandsSeparator: ' ',
+  decimalSeparator: '-',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const LAK = {
+  code: 'LAK',
+  symbol: '₭',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const LBP = {
+  code: 'LBP',
+  symbol: 'ل.ل.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const LKR = {
+  code: 'LKR',
+  symbol: '₨',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%s %ns%v',
+};
+
+export const LRD = {
+  code: 'LRD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const LSL = {
+  code: 'LSL',
+  symbol: 'M',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const LYD = {
+  code: 'LYD',
+  symbol: 'د.ل.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%ns%s%v',
+};
+
+export const MAD = {
+  code: 'MAD',
+  symbol: 'د.م.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const MDL = {
+  code: 'MDL',
+  symbol: 'lei',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const MGA = {
+  code: 'MGA',
+  symbol: 'Ar',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const MKD = {
+  code: 'MKD',
+  symbol: 'ден.',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const MMK = {
+  code: 'MMK',
+  symbol: 'K',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MNT = {
+  code: 'MNT',
+  symbol: '₮',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MOP = {
+  code: 'MOP',
+  symbol: 'MOP$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MRO = {
+  code: 'MRO',
+  symbol: 'UM',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const MTL = {
+  code: 'MTL',
+  symbol: '₤',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MUR = {
+  code: 'MUR',
+  symbol: '₨',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MVR = {
+  code: 'MVR',
+  symbol: 'MVR',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 1,
+  pattern: '%ns%v %s',
+};
+
+export const MWK = {
+  code: 'MWK',
+  symbol: 'MK',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MXN = {
+  code: 'MXN',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MYR = {
+  code: 'MYR',
+  symbol: 'RM',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const MZN = {
+  code: 'MZN',
+  symbol: 'MT',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%s%v',
+};
+
+export const NAD = {
+  code: 'NAD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const NGN = {
+  code: 'NGN',
+  symbol: '₦',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const NIO = {
+  code: 'NIO',
+  symbol: 'C$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const NOK = {
+  code: 'NOK',
+  symbol: 'kr',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const NPR = {
+  code: 'NPR',
+  symbol: '₨',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const NZD = {
+  code: 'NZD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const OMR = {
+  code: 'OMR',
+  symbol: '﷼',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%s %ns%v',
+};
+
+export const PAB = {
+  code: 'PAB',
+  symbol: 'B/.',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const PEN = {
+  code: 'PEN',
+  symbol: 'S/.',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const PGK = {
+  code: 'PGK',
+  symbol: 'K',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const PHP = {
+  code: 'PHP',
+  symbol: '₱',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const PKR = {
+  code: 'PKR',
+  symbol: '₨',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const PLN = {
+  code: 'PLN',
+  symbol: 'zł',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const PYG = {
+  code: 'PYG',
+  symbol: '₲',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const QAR = {
+  code: 'QAR',
+  symbol: '﷼',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const RON = {
+  code: 'RON',
+  symbol: 'lei',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const RSD = {
+  code: 'RSD',
+  symbol: 'Дин.',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const RUB = {
+  code: 'RUB',
+  symbol: '₽',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const RWF = {
+  code: 'RWF',
+  symbol: 'RWF',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const SAR = {
+  code: 'SAR',
+  symbol: '﷼',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const SBD = {
+  code: 'SBD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SCR = {
+  code: 'SCR',
+  symbol: '₨',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SDD = {
+  code: 'SDD',
+  symbol: 'LSd',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const SDG = {
+  code: 'SDG',
+  symbol: '£‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SEK = {
+  code: 'SEK',
+  symbol: 'kr',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const SGD = {
+  code: 'SGD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SHP = {
+  code: 'SHP',
+  symbol: '£',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SLL = {
+  code: 'SLL',
+  symbol: 'Le',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SOS = {
+  code: 'SOS',
+  symbol: 'S',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SRD = {
+  code: 'SRD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const STD = {
+  code: 'STD',
+  symbol: 'Db',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SVC = {
+  code: 'SVC',
+  symbol: '₡',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const SYP = {
+  code: 'SYP',
+  symbol: '£',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const SZL = {
+  code: 'SZL',
+  symbol: 'E',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const THB = {
+  code: 'THB',
+  symbol: '฿',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const TJS = {
+  code: 'TJS',
+  symbol: 'TJS',
+  thousandsSeparator: ' ',
+  decimalSeparator: ';',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const TMT = {
+  code: 'TMT',
+  symbol: 'm',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const TND = {
+  code: 'TND',
+  symbol: 'د.ت.‏',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 3,
+  pattern: '%s %ns%v',
+};
+
+export const TOP = {
+  code: 'TOP',
+  symbol: 'T$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const TRY = {
+  code: 'TRY',
+  symbol: 'TL',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const TTD = {
+  code: 'TTD',
+  symbol: 'TT$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const TVD = {
+  code: 'TVD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const TWD = {
+  code: 'TWD',
+  symbol: 'NT$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const TZS = {
+  code: 'TZS',
+  symbol: 'TSh',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const UAH = {
+  code: 'UAH',
+  symbol: '₴',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const UGX = {
+  code: 'UGX',
+  symbol: 'USh',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const USD = {
+  code: 'USD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const UYU = {
+  code: 'UYU',
+  symbol: '$U',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const UZS = {
+  code: 'UZS',
+  symbol: 'сўм',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v %s',
+};
+
+export const VEB = {
+  code: 'VEB',
+  symbol: 'Bs.',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const VEF = {
+  code: 'VEF',
+  symbol: 'Bs. F.',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const VND = {
+  code: 'VND',
+  symbol: '₫',
+  thousandsSeparator: '.',
+  decimalSeparator: ',',
+  decimalDigits: 1,
+  pattern: '%ns%v %s',
+};
+
+export const VUV = {
+  code: 'VUV',
+  symbol: 'VT',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 0,
+  pattern: '%ns%v%s',
+};
+
+export const WST = {
+  code: 'WST',
+  symbol: 'WS$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const XAF = {
+  code: 'XAF',
+  symbol: 'F',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const XCD = {
+  code: 'XCD',
+  symbol: '$',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const XOF = {
+  code: 'XOF',
+  symbol: 'F',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const XPF = {
+  code: 'XPF',
+  symbol: 'F',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%v%s',
+};
+
+export const YER = {
+  code: 'YER',
+  symbol: '﷼',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%s %ns%v',
+};
+
+export const ZAR = {
+  code: 'ZAR',
+  symbol: 'R',
+  thousandsSeparator: ' ',
+  decimalSeparator: ',',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const ZMW = {
+  code: 'ZMW',
+  symbol: 'ZK',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
+export const WON = {
+  code: 'WON',
+  symbol: '₩',
+  thousandsSeparator: ',',
+  decimalSeparator: '.',
+  decimalDigits: 2,
+  pattern: '%ns%s%v',
+};
+
